guidedog-api/requirements.txt
absl-py==0.13.0
astunparse==1.6.3
azure-cognitiveservices-vision-computervision==0.9.0
azure-common==1.1.27
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.5
clang==5.0
click==8.0.1
Flask==2.0.1
flatbuffers==1.12
gast==0.4.0
google-api-core==2.0.1
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-cloud-core==2.0.0
google-cloud-storage==1.42.2
google-cloud-texttospeech==2.5.2
google-cloud-vision==2.4.2
google-crc32c==1.1.2
google-pasta==0.2.0
google-resumable-media==2.0.2
googleapis-common-protos==1.53.0
grpcio==1.40.0
h5py==3.1.0
idna==3.2
isodate==0.6.0
itsdangerous==2.0.1
Jinja2==3.0.1
keras==2.6.0
Keras-Preprocessing==1.1.2
Markdown==3.3.4
MarkupSafe==2.0.1
msrest==0.6.21
numpy==1.19.5
oauthlib==3.1.1
opt-einsum==3.3.0
packaging==21.0
Pillow==8.3.2
proto-plus==1.19.0
protobuf==3.18.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==2.4.7
python-dotenv==0.19.0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.15.0
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.6.0
tensorflow-estimator==2.6.0
termcolor==1.1.0
typing-extensions==3.7.4.3
urllib3==1.26.6
Werkzeug==2.0.1
wrapt==1.12.1
